Refactor [#13] User 관련한 리뷰사항 반영 및 패키지 변경작업 진행
