+++ morib/src/main/java/org/morib/server/domain/user/application/UserManager.java
@@ -1,0 +1,5 @@
+package org.morib.server.domain.user.application;
+
+public class UserManager {
+
+}
